REFACTOR: moved new and deploy into separate functions
